requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.2.2=py35_0
astor=0.6.2=py35_0
blas=1.0=mkl
bleach=1.5.0=py35_0
ca-certificates=2018.4.16=0
certifi=2018.4.16=py35_0
freetype=2.8.1=vc14_0
gast=0.2.0=py35_0
grpcio=1.12.1=py35h1a1b453_0
html5lib=0.9999999=py35_0
icc_rt=2017.0.4=h97af966_0
icu=58.2=vc14_0
intel-openmp=2018.0.3=0
jpeg=9b=vc14_2
libpng=1.6.34=vc14_0
libprotobuf=3.5.2=vc14_0
libtiff=4.0.9=vc14_0
libwebp=0.5.2=vc14_7
markdown=2.6.11=py35_0
mkl=2018.0.3=1
mkl_fft=1.0.1=py35h452e1ab_0
mkl_random=1.0.1=py35h9258bd6_0
numpy=1.14.5=py35h9fa60d3_0
numpy-base=1.14.5=py35h5c71026_0
olefile=0.45.1=py35_0
opencv=3.4.1=py35_200
openssl=1.0.2o=vc14_0
pillow=5.1.0=py35h0738816_0
pip=10.0.1=py35_0
protobuf=3.5.2=py35_vc14_0
python=3.5.5=h0c2934d_2
qt=5.6.2=vc14_1
setuptools=39.2.0=py35_0
six=1.11.0=py35hc1da2df_1
tensorboard=1.8.0=py35he025d50_0
tensorflow=1.8.0=0
tensorflow-base=1.8.0=py35h1a1b453_0
termcolor=1.1.0=py35_1
tk=8.6.7=vc14_0
vc=14=h0510ff6_3
vs2015_runtime=14.0.25123=3
werkzeug=0.14.1=py35_0
wheel=0.31.1=py35_0
wincertstore=0.2=py35hfebbdb8_0
zlib=1.2.11=vc14_0
